mock-backend/pyproject.toml
[project]
name = "mock-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.117.1",
    "httpx>=0.28.1",
    "langchain-azure-dynamic-sessions>=0.2.0",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "py-orchestrate>=1.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.37.0",
    "azure-storage-blob>=12.19.0",
    "azure-search-documents>=11.4.0",
    "azure-core>=1.29.0",
    "openai>=1.12.0",
    "azure-ai-documentintelligence>=1.0.2",
    "langgraph-checkpoint-cosmosdb==0.2.4",
    "opentelemetry-instrumentation-langchain>=0.48.1",
    "langchain-azure-ai==0.1.8",
    "azure-monitor-opentelemetry==1.8.0",
]
